+++ Source/Diagnostics/ReducedDiags/FieldProbeParticleContainer.cpp
@@ -64,7 +64,7 @@
 using namespace amrex;
 
 FieldProbeParticleContainer::FieldProbeParticleContainer (AmrCore* amr_core)
-	:ParticleContainer<0,0,static_cast<int>(ParticleVal::nattribs)>(amr_core->GetParGDB())
+	:ParticleContainer<0,0,ParticleVal::nattribs>(amr_core->GetParGDB())
 {
 	SetParticleSize();
 }
@@ -114,13 +114,13 @@ FieldProbeParticleContainer::AddNParticles (int /*lev*/,
         //write Real attributes (SoA) to particle initialized zero
         DefineAndReturnParticleTile(0, 0, 0);
 
-		pinned_tile.push_back_real(static_cast<int>(ParticleVal::Ex), np, 0.0);
-		pinned_tile.push_back_real(static_cast<int>(ParticleVal::Ey), np, 0.0);
-		pinned_tile.push_back_real(static_cast<int>(ParticleVal::Ez), np, 0.0);
-		pinned_tile.push_back_real(static_cast<int>(ParticleVal::Bx), np, 0.0);
-		pinned_tile.push_back_real(static_cast<int>(ParticleVal::By), np, 0.0);
-		pinned_tile.push_back_real(static_cast<int>(ParticleVal::Bz), np, 0.0);
-		pinned_tile.push_back_real(static_cast<int>(ParticleVal::S), np, 0.0);
+		pinned_tile.push_back_real(ParticleVal::Ex, np, 0.0);
+		pinned_tile.push_back_real(ParticleVal::Ey, np, 0.0);
+		pinned_tile.push_back_real(ParticleVal::Ez, np, 0.0);
+		pinned_tile.push_back_real(ParticleVal::Bx, np, 0.0);
+		pinned_tile.push_back_real(ParticleVal::By, np, 0.0);
+		pinned_tile.push_back_real(ParticleVal::Bz, np, 0.0);
+		pinned_tile.push_back_real(ParticleVal::S, np, 0.0);
 	}
 
     /*
